heading_extraction/requirements.txt
opencv-python
pdf2image
google-cloud-vision
google-auth
pandas
python-dotenv
